+++ extensions/spotify-player/src/api/getMySavedShows.ts
@@ -1,24 +1,23 @@
 import { getErrorMessage } from "../helpers/getError";
-import { SimplifiedShowObject } from "../helpers/spotify.api";
+import { SimplifiedShowObject, SavedShowObject } from "../helpers/spotify.api";
 import { getSpotifyClient } from "../helpers/withSpotifyClient";
+import { iterate } from "../helpers/spotifyIterator";
 
-type GetMySavedShowsProps = { limit?: number };
+type GetMySavedShowsProps = { limit: number };
 
-export async function getMySavedShows({ limit = 50 }: GetMySavedShowsProps = {}) {
+export async function getMySavedShows({ limit }: GetMySavedShowsProps) {
   const { spotifyClient } = getSpotifyClient();
-
+  const iterator = iterate<SavedShowObject>(limit, (input) => spotifyClient.getMeShows(input));
   try {
-    const response = await spotifyClient.getMeShows({ limit });
-
-    // Normalize the response to match the SimplifiedShowObject type
-    // because the Spotify API returns a SavedShowObject type
-    const shows = (response?.items ?? []).map((showItem) => {
-      return {
-        ...showItem.show,
-      };
-    });
-
-    return { items: shows as SimplifiedShowObject[] };
+    const shows: SimplifiedShowObject[] = [];
+    for await (const items of iterator) {
+      for (const showItem of items) {
+        shows.push({
+          ...showItem.show,
+        } as SimplifiedShowObject);
+      }
+    }
+    return { items: shows };
   } catch (err) {
     const error = getErrorMessage(err);
     console.log("getMySavedShows.ts Error:", error);