+++ extensions/spotify-player/src/api/getMySavedTracks.ts
@@ -1,24 +1,23 @@
 import { getErrorMessage } from "../helpers/getError";
-import { SimplifiedTrackObject } from "../helpers/spotify.api";
+import { SimplifiedTrackObject, SavedTrackObject } from "../helpers/spotify.api";
 import { getSpotifyClient } from "../helpers/withSpotifyClient";
+import { iterate } from "../helpers/spotifyIterator";
 
-type GetMySavedTracksProps = { limit?: number };
+type GetMySavedTracksProps = { limit: number };
 
-export async function getMySavedTracks({ limit = 50 }: GetMySavedTracksProps = {}) {
+export async function getMySavedTracks({ limit }: GetMySavedTracksProps) {
   const { spotifyClient } = getSpotifyClient();
-
+  const iterator = iterate<SavedTrackObject>(limit, (input) => spotifyClient.getMeTracks(input));
   try {
-    const response = await spotifyClient.getMeTracks({ limit });
-
-    // Normalize the response to match the SimplifiedTrackObject type
-    // because the Spotify API returns a SavedTrackObject type
-    const tracks = (response?.items ?? []).map((trackItem) => {
-      return {
-        ...trackItem.track,
-      };
-    });
-
-    return { items: tracks as SimplifiedTrackObject[], total: response?.total ?? 0 };
+    const episodes: SimplifiedTrackObject[] = [];
+    for await (const items of iterator) {
+      for (const showItem of items) {
+        episodes.push({
+          ...showItem.track,
+        } as SimplifiedTrackObject);
+      }
+    }
+    return { items: episodes, total: episodes.length };
   } catch (err) {
     const error = getErrorMessage(err);
     console.log("getMySavedTracks.ts Error:", error);